requirements.txt
easyocr
fastapi
uvicorn
langchain-ollama
python-multipart
python-dotenv
sqlalchemy
bcrypt
psycopg2